iteration3 final

+++ src/main/java/ru/javacourse/eventmanagement/service/LocationService.java
@@ -11,7 +11,7 @@ import org.springframework.validation.annotation.Validated;
 import ru.javacourse.eventmanagement.db.entity.location.LocationEntity;
 import ru.javacourse.eventmanagement.db.repository.LocationRepository;
 import ru.javacourse.eventmanagement.domain.locations.Location;
-import ru.javacourse.eventmanagement.domain.locations.LocationMapper;
+import ru.javacourse.eventmanagement.domain.mapper.LocationMapper;
 
 import java.util.List;
 
@@ -24,18 +24,19 @@ public class LocationService {
     private final LocationRepository locationRepository;
     private final LocationMapper locationMapper;
 
+
     public List<Location> findAll() {
         return locationRepository.findAll()
                 .stream()
-                .map(locationMapper::mapFromEntity)
+                .map(locationMapper::mapFromEntityToLocation)
                 .toList();
     }
 
     @Transactional
     public Location createNewLocation(@Valid @NotNull Location location) {
-        var locationEntity = locationMapper.mapToEntity(location);
+        var locationEntity = locationMapper.mapFromLocationToEntity(location);
         locationEntity = locationRepository.save(locationEntity);
-        return locationMapper.mapFromEntity(locationEntity);
+        return locationMapper.mapFromEntityToLocation(locationEntity);
     }
 
     @Transactional
@@ -43,14 +44,18 @@ public class LocationService {
                                    @NotNull
                                    Long locationId) {
         var locationEntity = getLocationEntityById(locationId);
-        locationRepository.deleteById(locationId);
-        return locationMapper.mapFromEntity(locationEntity);
+        if (locationEntity.getEvents().isEmpty()) {
+            locationRepository.deleteById(locationId);
+        } else {
+            throw new IllegalArgumentException("It is impossible to delete a location because it contains events");
+        }
+        return locationMapper.mapFromEntityToLocation(locationEntity);
     }
 
     public Location getLocationByID(@PositiveOrZero(message = "ID must be positive or zero")
                                     @NotNull
                                     Long locationId) {
-        return locationMapper.mapFromEntity(getLocationEntityById(locationId));
+        return locationMapper.mapFromEntityToLocation(getLocationEntityById(locationId));
     }
 
     private LocationEntity getLocationEntityById(@NotNull Long locationId) {
@@ -72,7 +77,7 @@ public class LocationService {
         locationEntity.setCapacity(location.capacity());
         locationEntity.setDescription(location.description());
         var updateLocation = locationRepository.save(locationEntity);
-        return locationMapper.mapFromEntity(updateLocation);
+        return locationMapper.mapFromEntityToLocation(updateLocation);
     }
 
     private void isChangingCapacityCorrect(@NotNull Location locationUpdate, @NotNull LocationEntity locationByIdOld) {
